requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
dj-database-url==0.5.0
Django==3.2.5
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
